Rename efi -> elf
