+++ frontend/src/elf/contracts.ts
@@ -1,5 +1,5 @@
 import { ERC20Permit__factory } from "elf-council-typechain";
-import { defaultProvider } from "src/efi/providers/providers";
+import { defaultProvider } from "src/elf/providers/providers";
 import tokenList from "src/tokenlist.json";
 
 export const elementTokenContract = ERC20Permit__factory.connect(
